Added Swedish language
I also found a typo in the source string `Maximum Cuncurrent Downloads`, it should be `Concurrent`

Kind of a typo `Path To Cookie FIle` because the letter `i` is in upper case.

Also a kind of typo `Engine &quot;%1&quot; requires atleast version &quot;%2&quot; of Media Downloader`, `at least` should be two words.

Systray or Tray is now called `Notification area`, so maybe you should change `Show Tray Icon` to `Show icon in Notification area`?

If i understand this correct you need to add a new line to every language file when a new language will be added? Maybe you should think of making that a bit easier?

I was thinking of sorting the languages in alphabetical order, but that require changes in all the language file, so i won't do it (maybe you can write some code that will sort them in the dropdown menu?).