Make "Windows Only" proxy option disappear on non-Windows
Instead of only disabling the radio option "Use System Proxy(Windows Only)" if not on Windows, hides it and moves the other options up to fill its place at runtime.
It removes the unnecessary button on non-Windows and removes the clutter of saying it's Windows only when you're actually on Windows.
Additionally, disabling a radio button doesn't show any difference on my theming so you couldn't even tell it was disabled unless you clicked on it.